refactor(api): extract backup_runner + docs router from route modules (R-18, R-12)
R-18 (CA-02): move backup orchestration (`_process_one_device` + the
renamed public `run_backup_job`, was `_run_backup_job`) out of
`api/routes/backups.py` into `services/backup_runner.py` — a sibling of
`services/migration_pipeline.py`, so the scheduler no longer reaches
*into* a route module to borrow the backup engine. The documented SSH
mock-point `netcanon.api.routes.backups.get_collector` is preserved: the
service resolves the factory back through the route module via a
function-local import at call time, so the ~60 tests that patch that
string need zero edits. `backups.py` keeps `get_collector` imported
(F401 seam) + `BaseConfigStore` (still a `create_backup` type hint). The
lazy import keeps the route<->service dependency acyclic.

R-12 (CE-01): move the ~440-line `/docs` Swagger dark-mode reskin (six
`_DOCS_*` constants + the `swagger_ui` handler) out of
`api/routes/ui.py` into a new `api/routes/docs.py` with its own
`include_in_schema=False` router, registered at root in `main.py` so the
`/docs` URL is byte-for-byte unchanged (`docs_url=None` stays). `ui.py`
keeps the eight HTML page handlers + `HTMLResponse`; only
`get_swagger_ui_html` was removed there.

Both behaviour-preserving. Validated: full unit suite + tests/integration
(incl. the location-agnostic 24-method test_swagger_docs_page.py
contract) + scheduler e2e all green. New guard test pins that the
`_DOCS_*` constants moved (gone from ui.py, present in docs.py) and that
all root pages still resolve. De-staled doc refs in `_errors.py` +
`schedules.py` docstrings. Actuated from result-RA-18.md + result-RA-12.md.
